# Compression tables properly *used* for Outgoing Compression