--- a/akka-remote-tests/src/multi-jvm/scala/akka/remote/artery/MaxThroughputSpec.scala
+++ b/akka-remote-tests/src/multi-jvm/scala/akka/remote/artery/MaxThroughputSpec.scala
@@ -6,9 +6,11 @@ package akka.remote.artery
 import java.nio.ByteBuffer
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit.NANOSECONDS
+
 import scala.concurrent.duration._
 import akka.actor._
 import akka.remote.RemoteActorRefProvider
+import akka.remote.artery.compress.CompressionSettings
 import akka.remote.testconductor.RoleName
 import akka.remote.testkit.MultiNodeConfig
 import akka.remote.testkit.MultiNodeSpec
@@ -28,7 +30,7 @@ object MaxThroughputSpec extends MultiNodeConfig {
   commonConfig(debugConfig(on = false).withFallback(
     ConfigFactory.parseString(s"""
        # for serious measurements you should increase the totalMessagesFactor (20)
-       akka.test.MaxThroughputSpec.totalMessagesFactor = 20.0
+       akka.test.MaxThroughputSpec.totalMessagesFactor = 1.0
        akka {
          loglevel = ERROR
          # avoid TestEventListener
@@ -56,10 +58,10 @@ object MaxThroughputSpec extends MultiNodeConfig {
            #advanced.aeron-dir = "target/aeron"
 
            advanced.compression {
-             enabled = on
+             enabled = off
              actor-refs {
                enabled = on
-               advertisement-interval = 3 seconds
+               advertisement-interval = 1 second
              }
            }
          }
@@ -139,11 +141,14 @@ object MaxThroughputSpec extends MultiNodeConfig {
         println(
           s"=== MaxThroughput ${self.path.name}: " +
             f"throughput ${throughput * testSettings.senderReceiverPairs}%,.0f msg/s, " +
-            f"${throughput * payloadSize * testSettings.senderReceiverPairs}%,.0f bytes/s, " +
+            f"${throughput * payloadSize * testSettings.senderReceiverPairs}%,.0f bytes/s (payload), " +
+            f"${throughput * totalSize(context.system) * testSettings.senderReceiverPairs}%,.0f bytes/s (total" +
+            (if (CompressionSettings(context.system).enabled) ",compression" else "") + "), " +
             s"dropped ${totalMessages - totalReceived}, " +
             s"max round-trip $maxRoundTripMillis ms, " +
             s"burst size $burstSize, " +
             s"payload size $payloadSize, " +
+            s"total size ${totalSize(context.system)}, " +
             s"$took ms to deliver $totalReceived messages")
         plotRef ! PlotResult().add(testName, throughput * payloadSize * testSettings.senderReceiverPairs / 1024 / 1024)
         context.stop(self)
@@ -173,7 +178,10 @@ object MaxThroughputSpec extends MultiNodeConfig {
     totalMessages:       Long,
     burstSize:           Int,
     payloadSize:         Int,
-    senderReceiverPairs: Int)
+    senderReceiverPairs: Int) {
+    // data based on measurement
+    def totalSize(system: ActorSystem) = payloadSize + (if (CompressionSettings(system).enabled) 38 else 110)
+  }
 
   class TestSerializer(val system: ExtendedActorSystem) extends SerializerWithStringManifest with ByteBufferSerializer {
 
@@ -254,25 +262,25 @@ abstract class MaxThroughputSpec
       totalMessages = adjustedTotalMessages(20000),
       burstSize = 1000,
       payloadSize = 100,
-      senderReceiverPairs = 1) //,
-  //    TestSettings(
-  //      testName = "1-to-1-size-1k",
-  //      totalMessages = adjustedTotalMessages(20000),
-  //      burstSize = 1000,
-  //      payloadSize = 1000,
-  //      senderReceiverPairs = 1),
-  //    TestSettings(
-  //      testName = "1-to-1-size-10k",
-  //      totalMessages = adjustedTotalMessages(10000),
-  //      burstSize = 1000,
-  //      payloadSize = 10000,
-  //      senderReceiverPairs = 1),
-  //    TestSettings(
-  //      testName = "5-to-5",
-  //      totalMessages = adjustedTotalMessages(20000),
-  //      burstSize = 200, // don't exceed the send queue capacity 200*5*3=3000
-  //      payloadSize = 100,
-  //      senderReceiverPairs = 5)
+      senderReceiverPairs = 1),
+    TestSettings(
+      testName = "1-to-1-size-1k",
+      totalMessages = adjustedTotalMessages(20000),
+      burstSize = 1000,
+      payloadSize = 1000,
+      senderReceiverPairs = 1),
+    TestSettings(
+      testName = "1-to-1-size-10k",
+      totalMessages = adjustedTotalMessages(10000),
+      burstSize = 1000,
+      payloadSize = 10000,
+      senderReceiverPairs = 1),
+    TestSettings(
+      testName = "5-to-5",
+      totalMessages = adjustedTotalMessages(20000),
+      burstSize = 200, // don't exceed the send queue capacity 200*5*3=3000
+      payloadSize = 100,
+      senderReceiverPairs = 5)
   )
 
   def test(testSettings: TestSettings): Unit = {