… akka-bench-jmh/src/main/scala/akka/remote/artery/CodecBenchmark.scala
… akka-bench-jmh/src/main/scala/akka/remote/artery/compress/InvertCompressionTableBenchmark.scala
… akka-remote/src/main/scala/akka/remote/artery/Association.scala
… akka-remote/src/main/scala/akka/remote/artery/BufferPool.scala
… akka-remote/src/main/scala/akka/remote/artery/NoLiteralCompression.scala
… akka-remote/src/main/scala/akka/remote/artery/compress/ActualCompressions.scala
… akka-remote/src/main/scala/akka/remote/artery/compress/AdvertiseCompressionId.scala
… akka-remote/src/main/scala/akka/remote/artery/compress/AllCompressions.scala
… akka-remote/src/main/scala/akka/remote/artery/compress/CompressionTable.scala
… akka-remote/src/main/scala/akka/remote/artery/compress/DecompressionTable.scala
… akka-remote/src/main/scala/akka/remote/artery/compress/InboundCompressions.scala
… akka-remote/src/main/scala/akka/remote/artery/compress/OutboundActorRefCompression.scala
… akka-remote/src/test/scala/akka/remote/artery/EnvelopeBufferSpec.scala
… akka-remote/src/test/scala/akka/remote/artery/FlightRecorderSpec.scala
… akka-remote/src/test/scala/akka/remote/artery/compress/HandshakeShouldDropCompressionTableSpec.scala