Last week's PR finished the "advertise the table" part, however missed on actually using the tables for performing the outgoing writes. This complicates things a bit, however now the process is correct and we actually use the 2 table mode for compressing outgoing messages.

Lovely illustration of this below (`00000...` is the payload – 100 bytes):

```
[JVM-1] --------------------------------------------------------------------------------
[JVM-1]    ByteString(210 bytes)
[JVM-1]
[JVM-1]      00 00 00 00 C0 66 E1 B6 FF FF FF FF 04 00 00 00  | .....f..........
[JVM-1]      00 00 00 FF 24 00 00 00 6C 00 00 00 00 00 00 FF  | ....$...l.......
[JVM-1]      00 00 00 FF 46 00 61 72 74 65 72 79 3A 2F 2F 4D  | ....F.artery://M
[JVM-1]      61 78 54 68 72 6F 75 67 68 70 75 74 53 70 65 63  | axThroughputSpec
[JVM-1]      40 6C 6F 63 61 6C 68 6F 73 74 3A 35 30 32 36 32  | @localhost:50262
[JVM-1]      2F 75 73 65 72 2F 77 61 72 6D 75 70 2D 72 63 76  | /user/warmup-rcv
[JVM-1]      31 23 31 36 33 31 32 39 39 38 34 33 00 00 30 30  | 1#1631299843..00
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30                                            | 00
[JVM-1] --------Compression table gets applied! --------------------------------------
[JVM-1]    ByteString(138 bytes)
[JVM-1]
[JVM-1]      00 00 00 00 C0 66 E1 B6 FF FF FF FF 04 00 00 00  | .....f..........
[JVM-1]      00 00 00 FF 03 00 00 FF 24 00 00 00 01 00 00 FF  | ........$.......
[JVM-1]      00 00 00 FF 00 00 30 30 30 30 30 30 30 30 30 30  | ......0000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30 30 30 30 30 30 30  | 0000000000000000
[JVM-1]      30 30 30 30 30 30 30 30 30 30                    | 0000000000
[JVM-1] --------------------------------------------------------------------------------
```

I'm also playing around with aggressively optimising the hashing that's needed for updating the CMS structure. I think we may be able to get away with some sneaky reuse and internal knowledge what we're hashing (reusing arrays already, optimising for ActorRef.hashCode being fast etc). These changes may hurt preciseness, however that likely does not hurt us too much - will have to measure though.

Needs more tuning still, however we're on the right path slowly (the last commit is aggressive, but I think such things will save us time here):

Without compression:

```
[JVM-2] 1-to-1: 106,653 msgs/sec, 10,665,338 bytes/sec, totals 196,181 messages 18 MB
[JVM-2] 1-to-1: 113,711 msgs/sec, 11,371,114 bytes/sec, totals 310,025 messages 29 MB
[JVM-2] 1-to-1: 110,449 msgs/sec, 11,044,860 bytes/sec, totals 420,888 messages 40 MB
[JVM-2] 1-to-1: 112,424 msgs/sec, 11,242,395 bytes/sec, totals 533,871 messages 50 MB
[JVM-2] 1-to-1: 114,562 msgs/sec, 11,456,202 bytes/sec, totals 649,025 messages 61 MB
[JVM-2] 1-to-1: 115,744 msgs/sec, 11,574,440 bytes/sec, totals 765,179 messages 72 MB
[JVM-2] 1-to-1: 120,723 msgs/sec, 12,072,286 bytes/sec, totals 886,213 messages 84 MB
[JVM-2] 1-to-1: 118,723 msgs/sec, 11,872,295 bytes/sec, totals 1,005,411 messages 95 MB
[JVM-2] 1-to-1: 112,223 msgs/sec, 11,222,295 bytes/sec, totals 1,118,000 messages 106 MB
[JVM-2] 1-to-1: 119,715 msgs/sec, 11,971,522 bytes/sec, totals 1,237,870 messages 118 MB
[JVM-2] 1-to-1: 117,430 msgs/sec, 11,742,955 bytes/sec, totals 1,355,819 messages 129 MB
[JVM-2] 1-to-1: 116,757 msgs/sec, 11,675,703 bytes/sec, totals 1,472,859 messages 140 MB
[JVM-2] 1-to-1: 110,895 msgs/sec, 11,089,527 bytes/sec, totals 1,583,880 messages 151 MB

[JVM-1] === MaxThroughput 1-to-1-snd1: throughput 114,761 msg/s, 11,476,115 bytes/s (payload), 24,099,842 bytes/s (total), dropped 0, max round-trip 90 ms, burst size 1000, payload size 100, total size 210, 13942 ms to deliver 1600000 messages
```

With compression, advertisement every 1 second:

```
[JVM-2] 1-to-1: 89,678 msgs/sec, 8,967,770 bytes/sec, totals 105,540 messages 10 MB
[JVM-2] 1-to-1: 110,176 msgs/sec, 11,017,625 bytes/sec, totals 216,260 messages 20 MB
[JVM-2] 1-to-1: 119,761 msgs/sec, 11,976,128 bytes/sec, totals 336,624 messages 32 MB
[JVM-2] 1-to-1: 111,312 msgs/sec, 11,131,237 bytes/sec, totals 448,562 messages 42 MB
[JVM-2] 1-to-1: 103,000 msgs/sec, 10,300,045 bytes/sec, totals 552,022 messages 52 MB
[JVM-2] 1-to-1: 99,439 msgs/sec, 9,943,903 bytes/sec, totals 652,015 messages 62 MB
[JVM-2] 1-to-1: 110,348 msgs/sec, 11,034,844 bytes/sec, totals 762,678 messages 72 MB
[JVM-2] 1-to-1: 112,088 msgs/sec, 11,208,826 bytes/sec, totals 875,331 messages 83 MB
[JVM-2] 1-to-1: 104,869 msgs/sec, 10,486,904 bytes/sec, totals 980,632 messages 93 MB
[JVM-2] 1-to-1: 102,300 msgs/sec, 10,230,001 bytes/sec, totals 1,083,529 messages 103 MB
[JVM-2] 1-to-1: 99,106 msgs/sec, 9,910,614 bytes/sec, totals 1,183,077 messages 112 MB
[JVM-2] 1-to-1: 105,057 msgs/sec, 10,505,690 bytes/sec, totals 1,288,492 messages 122 MB
[JVM-2] 1-to-1: 108,996 msgs/sec, 10,899,586 bytes/sec, totals 1,397,802 messages 133 MB
[JVM-2] 1-to-1: 107,889 msgs/sec, 10,788,897 bytes/sec, totals 1,506,311 messages 143 MB

[JVM-1] === MaxThroughput 1-to-1-snd1: throughput 106,475 msg/s, 10,647,501 bytes/s (payload), 14,693,552 bytes/s (total,compression), dropped 0, max round-trip 86 ms, burst size 1000, payload size 100, total size 138, 15027 ms to deliver 1600000 messages
```

(That's just on my laptop for now, let's properly benchmark soon)
I added a (payload) and (total) size, based on observations (it's "hard coded"), not sure if helpful.
Need to fix when we start measuring, i.e. after compression kicks in.

Also, when thinking about it I think we could bring back the LRU Cache anyway, since it would ease out the difference between "very active recently but not compressed yet" and the compressed version of a ref. It's just a small memory cost per connection and likely makes sense to bring back, WDYT