@patriknw 2016-07-04
LGTM, thanks for quick fix