@patriknw 2016-07-04
Superseded by https://github.com/akka/akka/pull/20870